+++ CHANGELOG.md
@@ -20,6 +20,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## Changed
 
+- (breaking): `DefaultPusQueueHelper`: `seq_cnt_provider`, `pus_verificator`
+  and `default_pus_apid` (formerly `pus_apid`) do not have default values anymore
+  and need to be specified explicitely. 
 - (breaking): Renamed `tmtccmd.config.com.ComIfCfgBase` to `ComCfgBase`
 - (breaking): `tmtccmd.com.ComInterface`: Change `get_id` to `id` property.
 - (breaking): TCP (`tmtccmd.com.TcpSpacePacketsComIF`) and `tmtccmd.com.UdpComIF`: